made movement a part of player class and moved location of config file
