+++ app.js
@@ -2,7 +2,7 @@ const express = require('express')
 const path = require('path')
 const app = express()
 
-//app.use('/api', myApiRouter)
+// app.use('/api', myApiRouter)
 
 app.use('/static', express.static(path.join(__dirname, 'static')))
